+++ anton_potapov/graph.cpp
@@ -1,6 +1,49 @@
-#include <cassert>
 #include "graph.hpp"
+#include <algorithm>
+#include <cassert>
 #include <random>
+#include <utility>
+#include <vector>
+#include "graph_traverser.hpp"
+
+std::string get_edge_color_string(const EdgeColor& color) {
+  switch (color) {
+    case EdgeColor::Gray:
+      return "gray";
+    case EdgeColor::Green:
+      return "green";
+    case EdgeColor::Blue:
+      return "blue";
+    case EdgeColor::Yellow:
+      return "yellow";
+    case EdgeColor::Red:
+      return "red";
+  }
+}
+
+Vertex::Vertex(const VertexId& vertex_id) : id(vertex_id) {}
+
+Edge::Edge(const EdgeId& edge_id,
+           const VertexId& vertex1,
+           const VertexId& vertex2,
+           const EdgeColor& edge_color)
+    : vertex1_id(vertex1),
+      vertex2_id(vertex2),
+      color(edge_color),
+      id(edge_id) {}
+
+const std::set<EdgeId>& Vertex::connected_edges() const {
+  return connected_edges_;
+}
+
+bool Vertex::has_edge_id(const EdgeId& edge_id) const {
+  return connected_edges_.find(edge_id) != connected_edges_.end();
+}
+
+void Vertex::add_edge(const EdgeId& edge_id) {
+  assert(!has_edge_id(edge_id) && "edge that is to be added already exists");
+  connected_edges_.insert(edge_id);
+}
 
 Graph& Graph::operator=(Graph&& other_graph) {
   other_graph.update_vertices_depth();
@@ -19,6 +62,67 @@ Graph::Graph(Graph&& other_graph) {
   next_vertex_id_ = std::move(other_graph.next_vertex_id_);
   next_edge_id_ = std::move(other_graph.next_edge_id_);
   vertices_at_depth_ = std::move(other_graph.vertices_at_depth_);
+}
+
+Vertex Graph::get_vertex(const VertexId& id) const {
+  Vertex vertex_copy = vertices_.at(id);
+  vertex_copy.depth =
+      GraphTraverser::dynamic_bfs(*this, updated_depth_).at(vertex_copy.id);
+  return vertex_copy;
+}
+
+Vertex& Graph::get_vertex(const VertexId& id) {
+  update_vertices_depth();
+  return vertices_.at(id);
+}
+
+const Edge& Graph::get_edge(const EdgeId& id) const {
+  return edges_.at(id);
+}
+
+Edge& Graph::get_edge(const EdgeId& id) {
+  const auto& const_this = *this;
+  return const_cast<Edge&>(const_this.get_edge(id));
+}
+
+int Graph::max_depth() {
+  update_vertices_depth();
+  return vertices_at_depth_.size() - 1;
+}
+
+int Graph::max_depth() const {
+  const auto updated_depths =
+      GraphTraverser::dynamic_bfs(*this, updated_depth_);
+  int ans = 0;
+  for (const auto& [vertex_id, vertex_depth] : updated_depths) {
+    if (vertex_depth > ans) {
+      ans = vertex_depth;
+    }
+  }
+  return ans;
+}
+
+std::map<VertexId, Vertex> Graph::vertices() const {
+  auto vertices_updated_copy = vertices_;
+  const auto updated_depths =
+      GraphTraverser::dynamic_bfs(*this, updated_depth_);
+  for (auto& [vertex_id, vertex] : vertices_updated_copy) {
+    vertex.depth = updated_depths.at(vertex_id);
+  }
+  return vertices_updated_copy;
+}
+
+const std::map<VertexId, Vertex>& Graph::vertices() {
+  update_vertices_depth();
+  return vertices_;
+}
+
+const std::map<EdgeId, Edge>& Graph::edges() const {
+  return edges_;
+}
+
+const VertexId& Graph::get_root_vertex_id() const {
+  return vertices_.begin()->first;
 }
 
 const std::set<VertexId>& Graph::get_vertices_at_depth(int depth) {
@@ -83,7 +187,7 @@ VertexId Graph::add_vertex() {
 
 EdgeId Graph::add_edge(const VertexId& vertex1_id,
                        const VertexId& vertex2_id,
-                       const EdgeColor& edge_color = EdgeColor::Gray) {
+                       const EdgeColor& edge_color) {
   assert(is_vertex_exists(vertex1_id) && "Vertex 1 doesn't exist");
   assert(is_vertex_exists(vertex2_id) && "Vertex 2 doesn't exist");
   assert(!is_connected(vertex1_id, vertex2_id) && "Vertices already connected");
@@ -106,6 +210,13 @@ EdgeId Graph::add_edge(const VertexId& vertex1_id,
   }
 
   return new_edge_id;
+}
+
+VertexId Graph::get_next_vertex_id() {
+  return next_vertex_id_++;
+}
+EdgeId Graph::get_next_edge_id() {
+  return next_edge_id_++;
 }
 
 int Graph::get_new_depth(const VertexId& vertex1_id,
