+++ anton_potapov/graph_traverser.cpp
@@ -1,10 +1,12 @@
 #include <map>
+#include <set>
+#include <queue>
 
 #include "graph.hpp"
+#include "graph_traverser.hpp"
 
-const std::map<VertexId, int>& dynamic_bfs(const Graph& graph,
-                                           int start_depth) {
-  
+std::map<VertexId, int> GraphTraverser::dynamic_bfs(const Graph& graph,
+                                                           int start_depth) {
   std::map<VertexId, int> depths;
   std::queue<VertexId> bfs_queue;
   std::set<VertexId> used;
