+++ anton_potapov/graph_traverser.hpp
@@ -5,6 +5,6 @@
 
 class GraphTraverser {
  public:
-  static const std::map<VertexId, int>& dynamic_bfs(const Graph& graph,
-                                                    int start_depth);
+  static std::map<VertexId, int> dynamic_bfs(const Graph& graph,
+                                             int start_depth);
 };
